refactor(Infrastructure): сгруппировал классы в Infrastructure
Meta: #94956
