+++ backend/src/Infrastructure/EventSubscriber/ExceptionEventSubscriber.php
@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Infrastructure\ApiException;
+namespace App\Infrastructure\EventSubscriber;
 
+use App\Infrastructure\ApiException\ApiException;
+use App\Infrastructure\ApiException\CreateExceptionJsonResponse;
 use App\Infrastructure\AsService;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
