+++ backend/src/Infrastructure/EventSubscriber/ResponseEventSubscriber.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Infrastructure;
+namespace App\Infrastructure\EventSubscriber;
 
+use App\Infrastructure\AsService;
 use RuntimeException;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
