+++ backend/src/Infrastructure/ValueObject/Email.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Infrastructure;
+namespace App\Infrastructure\ValueObject;
 
 use Doctrine\ORM\Mapping as ORM;
 use Webmozart\Assert\Assert;
