+++ backend/src/Infrastructure/ValueObject/ValueObject.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Infrastructure;
+namespace App\Infrastructure\ValueObject;
 
 /**
  * Интерфейс для всех ValueObject
