+++ backend/src/Task/Domain/TaskName.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Task\Domain;
 
-use App\Infrastructure\ValueObject;
+use App\Infrastructure\ValueObject\ValueObject;
 use Doctrine\ORM\Mapping as ORM;
 use Webmozart\Assert\Assert;
