+++ backend/src/User/Profile/Command/SaveProfile/CreateProfile.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\User\Profile\Command\SaveProfile;
 
 use App\Infrastructure\AsService;
-use App\Infrastructure\Phone;
+use App\Infrastructure\ValueObject\Phone;
 use App\User\Profile\Domain\Profile;
 use App\User\Profile\Domain\ProfileId;
 use App\User\Profile\Domain\Profiles;
