+++ backend/src/User/Profile/Domain/Profile.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\User\Profile\Domain;
 
-use App\Infrastructure\Phone;
+use App\Infrastructure\ValueObject\Phone;
 use App\User\SignUp\Domain\UserId;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Uid\Uuid;
