+++ backend/src/User/Profile/Domain/ProfileId.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\User\Profile\Domain;
 
-use App\Infrastructure\ValueObject;
+use App\Infrastructure\ValueObject\ValueObject;
 use Symfony\Component\Uid\Uuid;
 use Symfony\Component\Uid\UuidV7;
