+++ backend/src/User/SignUp/Domain/UserPassword.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\User\SignUp\Domain;
 
-use App\Infrastructure\ValueObject;
+use App\Infrastructure\ValueObject\ValueObject;
 use Doctrine\ORM\Mapping as ORM;
 use Webmozart\Assert\Assert;
