+++ backend/tests/Unit/Infrastructure/EmailTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Tests\Unit\Infrastructure;
 
-use App\Infrastructure\Email;
+use App\Infrastructure\ValueObject\Email;
 use InvalidArgumentException;
 use PHPUnit\Framework\Attributes\TestDox;
 use PHPUnit\Framework\TestCase;
